requirements.txt
gTTS==2.5.4
discord.py==2.6.4
